Extract shared cycle detection into connectionValidation.ts
Deduplicates identical BFS cycle detection from audioUtils.ts and
midiUtils.ts into a single wouldCreateCycle() in connectionValidation.ts.
Updates all imports including test files.
